+++ Kernel/platform-linc80/config.h
@@ -23,7 +23,7 @@
 #define PROC_SIZE   32	  /* Memory needed per process (inc udata) */
 
 #define SWAPDEV     (swap_dev)	/* A variable for dynamic, or a device major/minor */
-extern unsigned int swap_dev;
+extern uint16_t swap_dev;
 #define SWAP_SIZE   0x41 	/* 32.5K in blocks (prog + udata) */
 #define SWAPBASE    0x8000	/* start at the base of user mem */
 #define SWAPTOP	    0x10000UL	/* Swap out program */
