+++ Kernel/platform-simple80/config.h
@@ -14,6 +14,7 @@
 #define CONFIG_LARGE_IO_DIRECT(x)	1
 /* One memory bank */
 #define CONFIG_BANKS	1
+/* FIXME: need to support non CTC uart ticking at 150Hz */
 #define TICKSPERSEC 10      /* Ticks per second */
 #define PROGBASE    0x0000  /* also data base */
 #define PROGLOAD    0x0100  /* also data base */
@@ -23,7 +24,7 @@
 #define PROC_SIZE   61	  /* Memory needed per process (inc udata) */
 
 #define SWAPDEV     (swap_dev)	/* A variable for dynamic, or a device major/minor */
-extern unsigned int swap_dev;
+extern uint16_t swap_dev;
 #define SWAP_SIZE   0x7A 	/* 61K in blocks (prog + udata) */
 #define SWAPBASE    0x0000	/* start at the base of user mem */
 #define SWAPTOP	    0xF000	/* Swap out program */
